Split the predefined Active Task modules into a separate file.
