+++ ResearchKit/Common/ORKOrderedTask_Private.h
@@ -66,9 +66,10 @@ FOUNDATION_EXPORT NSString *const ORKHeartRateRecorderIdentifier;
 
 FOUNDATION_EXPORT void ORKStepArrayAddStep(NSMutableArray *array, ORKStep *step);
 
-@interface ORKOrderedTask ()
+@interface ORKOrderedTask (ORKMakeTaskUtilities)
 
 + (ORKCompletionStep *)makeCompletionStep;
++ (NSDateComponentsFormatter *)textTimeFormatter;
 
 @end
