+++ ResearchKit/ResearchKit.h
@@ -56,6 +56,7 @@
 
 #import <ResearchKit/ORKTask.h>
 #import <ResearchKit/ORKOrderedTask.h>
+#import <ResearchKit/ORKOrderedTask+ORKPredefinedActiveTask.h>
 #import <ResearchKit/ORKNavigableOrderedTask.h>
 #import <ResearchKit/ORKStepNavigationRule.h>
